LayoutTests/imported/w3c/web-platform-tests/html/webappapis/dynamic-markup-insertion/document-write/module-tla-delayed-expected.txt

CONSOLE MESSAGE: SyntaxError: Unexpected identifier 'delay'

Harness Error (TIMEOUT), message = null

TIMEOUT document.write in an imported module Test timed out
